wip #37 refactor ase stko optimizer
I moved it into its own python file

also added comments elsewhere and cleaned up commented code
